Tinkering with AI Based Tools
Tinkering with Maket and Adobe Firefly. The quality for Maket is better as it is designed specifically for Architects leveraging the advances in AI. Whereby Adobe Firefly is geared for everyone. I have only been tinkering on a surface level tonight. This is the article that mentioned 'Maket AI' --> https://architizer.com/blog/practice/tools/top-ai-tools-for-architects-and-designers/ and I simply googled 'AI Architect'