01-20-2024
I made a model using cardboard, paper, tape and some elastic. The cardboard was from a box with fries in it & the oil seeped into the paper. | This model has a circular arch shaped roof and earthen materials can then be placed on the roof. It has 3 adjustable sun-roofs. There are two entrances. But the entrances also serve as a shade. They are envisioned to have solar panels on them. But the ratio for the weights has not been factored as the opening and closing of these retractable larger doors slash roofs, that has to happen safely. Have ideas for that. I have been thinking about a structure that is incredibly resilient and can be manufactured for $5000 American or Canadian (inflation adjusted). That will solve the homelessness issue and greatly help improve the quality of life for a lot of people around the world.