Giving credit where it is due
I don't remember who created which file. But I recall sharing the direct link on the shangrila-vhp wiki. So far, I have not made substantial modifications to any file that I have downloaded from the web. So placing ethics at the center of everything one does. Just like it would be unfair for someone else to leverage my work to claim it as their own. The same rule applies to me.